Refactor typedefs
