+++ db_builders/typedefs/manufacturer.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Manufacturer(BaseModel):
+    title: str
+    url: str
