+++ db_builders/typedefs/omniclass.py
@@ -1,16 +1,6 @@
 from pydantic import BaseModel
 
 
-class Manufacturer(BaseModel):
-    title: str
-    url: str
-
-
-class Parameter(BaseModel):
-    name: str
-    values: list[str]
-
-
 class Omniclass(BaseModel):
     """ A class to represent an Omniclass product.
 
@@ -21,10 +11,3 @@ class Omniclass(BaseModel):
 
     def __str__(self):
         return f"{self.number} {self.name}"
-
-
-class SearchResultItem(BaseModel):
-    """ This is used to parse the results from the Google custom search API """
-    title: str
-    link: str
-    snippet: str
