+++ db_builders/typedefs/parameter.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Parameter(BaseModel):
+    name: str
+    values: list[str]
