+++ db_builders/typedefs/search_result.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class SearchResultItem(BaseModel):
+    """ This is used to parse the results from the Google custom search API """
+    title: str
+    link: str
+    snippet: str
